+++ src/main/scala/code/bankconnectors/ObpJvmMappedConnector.scala
@@ -8,7 +8,7 @@ import java.util.{Date, Locale, Optional, UUID}
 
 import code.api.util.ErrorMessages._
 import code.accountholder.{AccountHolders, MapperAccountHolders}
-import code.api.util.{APIUtil, CallContext, ErrorMessages}
+import code.api.util._
 import code.api.v2_1_0.TransactionRequestCommonBodyJSON
 import code.bankconnectors.vMar2017.InboundAdapterInfoInternal
 import code.branches.Branches.{Branch, BranchT}
