+++ src/main/scala/code/bankconnectors/vJune2017/KafkaMappedConnector_vJune2017.scala
@@ -31,7 +31,7 @@ import code.api.ResourceDocs1_4_0.SwaggerDefinitionsJSON
 import code.api.cache.Caching
 import code.api.util.APIUtil.{MessageDoc, getSecondsCache, saveConnectorMetric}
 import code.api.util.ErrorMessages._
-import code.api.util.{APIUtil, ApiSession, CallContext, ErrorMessages}
+import code.api.util._
 import code.api.util.APIUtil._
 import code.api.v3_1_0.{AccountV310Json, CardObjectJson, CheckbookOrdersJson}
 import code.atms.Atms.{AtmId, AtmT}
