+++ src/main/scala/code/metrics/ConnectorMetricsProvider.scala
@@ -2,10 +2,9 @@ package code.metrics
 
 import java.util.{Calendar, Date}
 
-import code.api.util.APIUtil
-import code.bankconnectors.OBPQueryParam
+import code.api.util.{APIUtil, OBPQueryParam}
 import code.remotedata.RemotedataConnectorMetrics
-import net.liftweb.util.{Props, SimpleInjector}
+import net.liftweb.util.SimpleInjector
 
 object ConnectorMetricsProvider extends SimpleInjector {
