+++ src/main/scala/code/metrics/MappedMetrics.scala
@@ -1,19 +1,18 @@
 package code.metrics
 
-import java.sql.{PreparedStatement, Time, Timestamp}
+import java.sql.{PreparedStatement, Timestamp}
 import java.util.Date
 
-import code.api.util.APIUtil
 import code.api.util.ErrorMessages._
-import code.bankconnectors.{OBPImplementedByPartialFunction, _}
+import code.api.util._
 import code.util.Helper.MdcLoggable
 import code.util.{MappedUUID, UUIDString}
-import net.liftweb.common.{Box, Empty, Full}
+import net.liftweb.common.{Box, Full}
 import net.liftweb.mapper.{Index, _}
 import net.liftweb.util.Helpers.tryo
 
-import scala.concurrent.ExecutionContext.Implicits.global
 import scala.collection.immutable.List
+import scala.concurrent.ExecutionContext.Implicits.global
 import scala.concurrent.Future
 
 object MappedMetrics extends APIMetrics with MdcLoggable{
