+++ src/main/scala/code/remotedata/RemotedataMetrics.scala
@@ -4,10 +4,11 @@ import java.util.Date
 
 import akka.pattern.ask
 import code.actorsystem.ObpActorInit
-import code.bankconnectors.OBPQueryParam
+import code.api.util.OBPQueryParam
 import code.customer.{AmountOfMoneyTrait => _}
 import code.metrics._
 import net.liftweb.common.Box
+
 import scala.concurrent.Future
