+++ src/test/scala/code/api/v1_4_0/AtmsTest.scala
@@ -1,10 +1,10 @@
 package code.api.v1_4_0
 
-import code.api.v1_4_0.JSONFactory1_4_0.{AtmJson, AtmsJson}
 import code.api.util.APIUtil.OAuth._
+import code.api.util.OBPQueryParam
+import code.api.v1_4_0.JSONFactory1_4_0.{AtmJson, AtmsJson}
 import code.atms.Atms.{AtmId, AtmT}
 import code.atms.{Atms, AtmsProvider}
-import code.bankconnectors.OBPQueryParam
 import code.common.{AddressT, LicenseT, LocationT, MetaT}
 import code.model.BankId
 import code.setup.DefaultUsers
