+++ src/test/scala/code/atms/MappedAtmsProviderTest.scala
@@ -1,7 +1,7 @@
 package code.atms
 
+import code.api.util.OBPLimit
 import code.atms.Atms.AtmT
-import code.bankconnectors.OBPLimit
 import code.model.BankId
 import code.setup.ServerSetup
 import net.liftweb.mapper.By
