+++ src/test/scala/code/setup/TestConnectorSetup.scala
@@ -2,11 +2,10 @@ package code.setup
 
 import java.util.{Calendar, Date}
 
-import code.api.util.APIUtil
-import code.bankconnectors.{Connector, OBPLimit, OBPOffset}
+import code.api.util.{APIUtil, OBPLimit, OBPOffset}
+import code.bankconnectors.Connector
 import code.metadata.counterparties.CounterpartyTrait
 import code.model._
-import net.liftweb.util.Helpers._
 
 trait TestConnectorSetup {
